Test the LoggableListener with both an annotation and attribute reader
